[fix]트랙 저장 시에 유저 아이디가 반영이 안되는 문제 해결

+++ src/main/java/com/running/you_run/running/payload/request/RunningTrackStoreRequest.java
@@ -12,7 +12,7 @@ public record RunningTrackStoreRequest(
         int totalDistance,
         double rate,
         List<CoordinateDto> path,
-        int userId
+        long userId
 ) {
     public LineString createLineString() {
         Coordinate[] coords = path.stream()
